Added unit test to referee(), fixed bugs in referee()
Added an argument and return value for this referee(), public bool referee(string[,] currentBoard). Now in unit test, I can pass a board(from testcase) to this function and get a return value(win or not), this is more test-friendly.
After run the unit test, I found there was a bug in referee(), it was unable to check the win situation in diagonal, because the index should be able to start from 0, but my condition was greater than 0.
Separated gameEnd() from referee(). Now referee() is a pure checking function, relocated gameEnd() to gameStart(), if the game fulfills some of the criteria, gameEnd() will be called.